This PR includes various improvements and bug fixes:

- faster file reading: Instead of creating a new Scanner object on each file line, there is now one single BufferedReader only. This is much faster (test: 3s vs 6s).
- faster training: Switched from random to sequential array access and merged some for loops to improve performance drastically (test: 6s vs 16s).
- more accurate training: Til now, the learning rate and the neighborhood were reduced by some growing factor over all iterations. The problem: after a short time both became negative. For example, quite soon ~0.03 is reduced per iteration from the current learning rate. As the initial value was 0.05, we almost directly jump in a negative area. This is why the map then acts randomly as we, i.e., move the BMU away from the input vector instead of adjusting it towards it. The learning rate is now reduced linearly towards 0 over all iterations. The neighborhood is reduced exponentially, so it settles down quickly.

Again as these variables became negative quite soon, the map showed some undefined behavior:
<img src="https://user-images.githubusercontent.com/9994218/83644796-f49dd400-a5b1-11ea-8ac3-195071d4fcd0.jpg" width="150">
Running the same file but with the linear and exponential decreasing methods:
<img src="https://user-images.githubusercontent.com/9994218/83644804-f7002e00-a5b1-11ea-9ef6-775c3fde9594.jpg" width="150">

The map not only gets trained correctly now, but the time savings are quite huge, i.e., the old version takes around three times longer, even on a small test file (50k lines with 512 columns each).

It also fixes #1.