@mpas97 2020-06-03
A few typos have been eliminated too.